// common/enums/MoveServicesEnum.ts
export enum MoveServiceOrder {
  MOVING = 'moving',
  CLEANING = 'cleaning',
  CLEANING_AND_MOVING = 'cleaningAndMoving',
  SHOW_CLEANING = 'showCleaning',
  DISCOUNT = 'discount',
  PACKAGING = 'packaging',
  STORAGE = 'storage',
}

export enum MoveService {
  PACKAGING = 'packaging',
  MOVEHELP = 'movehelp',
  MOVECLEAN = 'moveclean',
}

export enum ExtendedMoveService {
  PACKAGING = 'packaging',
  MOVEHELP = 'movehelp',
  MOVECLEAN = 'moveclean',
  MOVEHELP_COMBINED = 'movehelpCombined',
}